refactor: split large ImageBankPage into smaller components
- Extract ConfirmDeleteModal, ImageMetadataUserDetails, and ImageUploadPreview components
- Create useImageBank custom hook for business logic separation
- Add exifUtils utility for EXIF data handling
- Reduce ImageBankPage from 678 to ~300 lines (56% reduction)
- Improve code organization following Single Responsibility Principle
- Maintain all existing functionality while improving maintainability
- All tests passing
