+++ src/components/imageBank/index.ts
@@ -1,0 +1,3 @@
+export { ConfirmDeleteModal } from './ConfirmDeleteModal';
+export { ImageMetadataUserDetails } from './ImageMetadataUserDetails';
+export { ImageUploadPreview } from './ImageUploadPreview';
